modificacion a las formulas de costos y modificacion al campo de valor en la migracion

+++ database/migrations/CalidadPcs/2019_08_27_222942_create_proceso_costos_table.php
@@ -15,7 +15,7 @@ class CreateProcesoCostosTable extends Migration
     {
         Schema::connection('calidadpcs')->create('TBL_Calidadpcs_proceso_costos', function (Blueprint $table) {
             $table->increments('PK_CPC_Id_Costo');
-            $table->integer('CPC_Valor')->unsigned();
+            $table->bigInteger('CPC_Valor')->unsigned();
             $table->integer('FK_CPC_Id_Formula')->unsigned();
             $table->foreign('FK_CPC_Id_Formula')->references('PK_CPCI_Id_Costos')->on('TBL_Calidadpcs_proceso_costos_informacion')->onDelete("cascade");
             $table->integer('FK_CPC_Id_Proyecto')->unsigned();
